Changes to rename macros.h file so there is no clash with other macros.h files.
